+++ CHANGELOG.md
@@ -4,12 +4,19 @@ All notable changes to this project will be documented in this
 file. This change log is based upon [keep-a-changelog]. This project adheres to
 [Semantic Versioning].
 
-## Unreleased
+## [Unreleased]
+
+## [0.1.0] - 2015-11-27
 
 ###Added
 
 - Basic gem project folder and file structure (bin, lib, test).
-- CHANGELOG, contain notable changes to this gem.
+- CHANGELOG, contains notable changes to this gem.
+- DEVCORE, contains motivation and development details.
+- Dungeon, a generator which randomly connects classes that include Traversable.
+- Traversable, a mixin needed for dungeon to connect rooms.
 
 [keep-a-changelog]: https://github.com/olivierlacan/keep-a-changelog
 [Semantic Versioning]: http://semver.org/
+[Unreleased]: https://github.com/mbrand12/t10-dungeon/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/mbrand12/t10-dungeon/compare/64a26c1...v0.1.0